requirements.txt
asyncpg~=0.30.0
aiogram~=3.22.0
APScheduler~=3.11.0
python-decouple~=3.8
pip~=25.2
attrs~=25.3.0
wheel~=0.37.1
idna~=3.10
multidict~=6.6.4
propcache~=0.3.2
pydantic~=2.11.9
aiofiles~=24.1.0
certifi~=2025.8.3
aiohttp~=3.12.15
yarl~=1.20.1
aiosignal~=1.4.0
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
setuptools~=65.5.0
sqlalchemy~=2.0.43
aiosqlite~=0.21.0